fix: Split BackendStatusContext to satisfy react-refresh lint rule
Separates context creation, provider component, and hook into individual
files so each file only exports one type of thing.
