+++ src/context/BackendStatusContext.tsx
@@ -1,46 +1,3 @@
-import { createContext, useContext, useEffect, useState } from 'react'
-import { useQuery, useQueryClient } from '@tanstack/react-query'
-import axios from 'axios'
-import client from '../api/client'
+import { createContext } from 'react'
 
-function isNetworkError(error: unknown): boolean {
-  return axios.isAxiosError(error) && !error.response
-}
-
-const BackendStatusContext = createContext<boolean>(false)
-
-export function BackendStatusProvider({ children }: { children: React.ReactNode }) {
-  const [isOffline, setIsOffline] = useState(false)
-  const queryClient = useQueryClient()
-
-  useQuery({
-    queryKey: ['__health__'],
-    queryFn: () => client.get('/products'),
-    refetchInterval: 15_000,
-    retry: 0,
-    staleTime: 0,
-  })
-
-  useEffect(() => {
-    const unsubscribe = queryClient.getQueryCache().subscribe((event) => {
-      if (event.type !== 'updated') return
-      const { state } = event.query
-      if (state.status === 'error' && isNetworkError(state.error)) {
-        setIsOffline(true)
-      } else if (state.status === 'success') {
-        setIsOffline(false)
-      }
-    })
-    return unsubscribe
-  }, [queryClient])
-
-  return (
-    <BackendStatusContext.Provider value={isOffline}>
-      {children}
-    </BackendStatusContext.Provider>
-  )
-}
-
-export function useBackendStatus(): boolean {
-  return useContext(BackendStatusContext)
-}
+export const BackendStatusContext = createContext<boolean>(false)
