+++ src/context/useBackendStatus.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import { BackendStatusContext } from './BackendStatusContext'
+
+export function useBackendStatus(): boolean {
+  return useContext(BackendStatusContext)
+}
